add float number

+++ snack-1/script.js
@@ -1,9 +1,9 @@
 // l'utente inserisce due numeri in successione, con due prompt
 // il software stampa il maggiore
 
-const first_number = parseInt(prompt("Inserisci il primo numero")); // ask the user for the first number
+const first_number = parseFloat(prompt("Inserisci il primo numero")); // ask the user for the first number
 
-const second_number = parseInt(prompt("Inserisci il secondo numero")); // ask the user for the second number 
+const second_number = parseFloat(prompt("Inserisci il secondo numero")); // ask the user for the second number 
 
 if (first_number > second_number) {
   console.log(first_number);
